requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
build==0.10.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1.2
django-author==1.2.0
django-celery-results==2.4.0
django-ckeditor==6.2.0
django-import-export==3.0.2
django-import-export-celery==1.3
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-social-share==2.2.1
et-xmlfile==1.1.0
kombu==5.2.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
packaging==21.3
Pillow==9.0.1
pip-tools==6.12.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dotenv==0.21.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
setuptools-git==1.2
six==1.16.0
sqlparse==0.4.2
tablib==3.3.0
tomli==2.0.1
tzdata==2022.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
